requirements.txt
pandas
numpy
matplotlib
seaborn
requests
python-dotenv
python-dateutil
pymongo
PyYAML
plotly==5.10.0
ipywidgets>=7.6
jupyterlab>=3
jupyter-dash
wget
xmltodict
shapely
haversine
scikit-learn
scipy
